User Backend: Extract route calculation into service
